feat: inject renderOptions to appContext
